Split document router unit tests into separate files
